refactor: 불필요한 코드 삭제 및 컴포넌트화
